requirements.txt
django-appconf==1.0.2
requests>=2.0.0
